requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
Flask==3.1.1
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.6.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
websockets==15.0.1
Werkzeug==3.1.3
